rename: 페이지네이션 버튼 이름 변경
